### 2019-11-28 #1920 @Kevin-Mattheus-Moerman
@Deech08 thanks for your submission. It looks like you have a BSD license but it is stowed away in a folder along with several other files. I recommend that you create a plain text LICENSE file in your main repository folder. This way GitHub will pick up on the license and render a license badge at the top. Also whedon (our journal system bot) and our reviewers will find it easier to find the license.
