### 2019-11-29 #1920 @Deech08
Thanks for agreeing to edit this @xuanxu - I just pushed an update to my whampy repo with the LICENSE.rst file in the main folder.
