### 2019-12-02 #1920 @xuanxu
:wave: @ngoldbaum would you be willing to review this submission for JOSS? I think your knowledge of Python and Astrophysics makes you a good fit for this submission.
