### 2019-12-08 #1940 @whedon
[REVIEW]: whampy: Python Package to Interact with, Visualize, and Analyze the Wisconsin H-Alpha Mapper - Sky Survey
**Submitting author:** @Deech08 (<a href="http://orcid.org/0000-0002-7955-7359">Dhanesh Krishnarao</a>)
**Repository:** <a href="https://github.com/Deech08/whampy" target ="_blank">https://github.com/Deech08/whampy</a>
**Version:** v1.0.1
