### 2019-12-09 Deech08/whampy#2 @Gabriel-p
FileNotFoundError
Hi there, I'm one of the reviewers of your package for JOSS (see: https://github.com/openjournals/joss-reviews/issues/1940)

The example in the README includes loading a survey with:

```
>>> from whampy.skySurvey import SkySurvey
>>> # Load the Survey
>>> survey = SkySurvey()
```

but this fails with:

    FileNotFoundError: [Errno 2] No such file or directory: '/home/gabriel/miniconda3/envs/py3/lib/python3.7/site-packages/whampy/data/wham-ss-DR1-v161116-170912.fits'

Is this file supposed to come with the installed package or is must the user download it from somewhere else?
