### 2019-12-09 Deech08/whampy#2 @Deech08
Uh oh - if the file isn't found its "supposed" to download it from a url. I will check what's happening!
