### 2019-12-09 416b152
Fix default survey file search
if local file not found - defaults to getting remote data
