### 2019-12-09 a0196b8
Update local load
downloads file if not found on first load.
