### 2019-12-09 Deech08/whampy#2 @Deech08
This should now be resolved in the latest version on github and the latest pip installable version  (v0.1.0a7). It will now auto download the survey data into the default directory (within the package) on first load of the survey.
