### 2019-12-09 Deech08/whampy#2 @Deech08
Just as a heads up - the TravisCI build is currently failing because a server is down associated with the dustmaps package - an optional dependency. I will restart the build when I notice it is up and running again!
