### 2019-12-12 Deech08/whampy#3 @HeloiseS
Truncated File
## Issue

Following warning was received after calling `survey=SkySurvey()`:

`WARNING: File may have been truncated: actual file length (32145408) is smaller than the expected size (91200960) [astropy.io.fits.file]`

And the following error followed:

`ValueError: cannot reshape array of size 17492 into shape (49640,)`

## Set up prior to issue
whampy was installed in a python 3.8 conda environment. The scripts on the README page was followed.

The first attempt at `survey=SkySurvey()` did not finish running in under 3 minutes and I interrupted the prompt. The error mentioned above then arose on the second attempt. I exited the python prompt and started it again and received the same error.

I guess the download does not restart if the file already exists. It is not immediately clear where that file is stored or how to clear it.

## Recommendations

The warning indicates this issue is controlled to some extent, yet python crashes non-the-less. If it is not expected that whampy will work if the file is incomplete, it could be worth prompting the user to ask to automatically clear the truncated file and start a fresh download.

-----
*This issue is part of the current JOSS-review of whampy ( [#1940](https://github.com/openjournals/joss-reviews/issues/1940) )*
