### 2019-12-12 Deech08/whampy#2 @HeloiseS
> Just as a heads up - the TravisCI build is currently failing because a server is down associated with the dustmaps package - an optional dependency. I will restart the build when I notice it is up and running again!

Travis is very slow sometimes :) As for me I'll only check your travis build at the end once both me and @Gabriel-p have checked everything we wanted to check and once you've had the ability to close our issues - so no worries
