### 2019-12-12 Deech08/whampy#4 @HeloiseS
Clear link to documentation
I struggled a little bit to find a link to your documentation in your README.

The WHAM link under the badges is broken and I wonder if that wasn't meant to link to the docs (regardless you should update the broken link).

I found [this](https://whampy.readthedocs.io/en/latest/?badge=latest#getting-started) by clicking the doc:passing badge (I'll go through it later).

I think this deserves a link of it's own and a clear header in your README :)

-----
*This issue is part of the current JOSS-review of whampy ( [#1940](https://github.com/openjournals/joss-reviews/issues/1940) )*
