### 2019-12-12 #1940 @HeloiseS
I got a quick question: do we expect there to be an explicit list of dependencies even if the package is pip installable?
