### 2019-12-12 Deech08/whampy#2 @Gabriel-p
Upgraded to `v0.1.0a7`, now it attempts to download the file but after a while fails with:

```
>>> survey = SkySurvey()
Downloading survey to /home/gabriel/miniconda3/envs/py3/lib/python3.7/site-packages/whampy/data/wham-ss-DR1-v161116-170912.fits
Traceback (most recent call last):
  File "/home/gabriel/miniconda3/envs/py3/lib/python3.7/site-packages/whampy/skySurvey.py", line 188, in __init__
    f = open(filename)
FileNotFoundError: [Errno 2] No such file or directory: '/home/gabriel/miniconda3/envs/py3/lib/python3.7/site-packages/whampy/data/wham-ss-DR1-v161116-170912.fits'

Traceback (most recent call last):
  File "/home/gabriel/miniconda3/envs/py3/lib/python3.7/urllib/request.py", line 1317, in do_open
    encode_chunked=req.has_header('Transfer-encoding'))
  File "/home/gabriel/miniconda3/envs/py3/lib/python3.7/http/client.py", line 1244, in request
    self._send_request(method, url, body, headers, encode_chunked)
  File "/home/gabriel/miniconda3/envs/py3/lib/python3.7/http/client.py", line 1290, in _send_request
    self.endheaders(body, encode_chunked=encode_chunked)
  File "/home/gabriel/miniconda3/envs/py3/lib/python3.7/http/client.py", line 1239, in endheaders
    self._send_output(message_body, encode_chunked=encode_chunked)
  File "/home/gabriel/miniconda3/envs/py3/lib/python3.7/http/client.py", line 1026, in _send_output
    self.send(msg)
  File "/home/gabriel/miniconda3/envs/py3/lib/python3.7/http/client.py", line 966, in send
    self.connect()
  File "/home/gabriel/miniconda3/envs/py3/lib/python3.7/http/client.py", line 938, in connect
    (self.host,self.port), self.timeout, self.source_address)
  File "/home/gabriel/miniconda3/envs/py3/lib/python3.7/socket.py", line 727, in create_connection
    raise err
  File "/home/gabriel/miniconda3/envs/py3/lib/python3.7/socket.py", line 716, in create_connection
    sock.connect(sa)
TimeoutError: [Errno 110] Connection timed out

Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/gabriel/miniconda3/envs/py3/lib/python3.7/site-packages/whampy/skySurvey.py", line 193, in __init__
    urllib.request.urlretrieve(url, filename)
  File "/home/gabriel/miniconda3/envs/py3/lib/python3.7/urllib/request.py", line 247, in urlretrieve
    with contextlib.closing(urlopen(url, data)) as fp:
  File "/home/gabriel/miniconda3/envs/py3/lib/python3.7/urllib/request.py", line 222, in urlopen
    return opener.open(url, data, timeout)
  File "/home/gabriel/miniconda3/envs/py3/lib/python3.7/urllib/request.py", line 525, in open
    response = self._open(req, data)
  File "/home/gabriel/miniconda3/envs/py3/lib/python3.7/urllib/request.py", line 543, in _open
    '_open', req)
  File "/home/gabriel/miniconda3/envs/py3/lib/python3.7/urllib/request.py", line 503, in _call_chain
    result = func(*args)
  File "/home/gabriel/miniconda3/envs/py3/lib/python3.7/urllib/request.py", line 1345, in http_open
    return self.do_open(http.client.HTTPConnection, req)
  File "/home/gabriel/miniconda3/envs/py3/lib/python3.7/urllib/request.py", line 1319, in do_open
    raise URLError(err)
urllib.error.URLError: <urlopen error [Errno 110] Connection timed out>
```

You should consider using a smaller file or adding it to the package.
