### 2019-12-12 Deech08/whampy#2 @HeloiseS
I had the same problem and then the following issue [occured](https://github.com/Deech08/whampy/issues/3)

It would be worth warning the user that this is a big file and will take a while to download if it's not added to the package.
