### 2019-12-12 Deech08/whampy#2 @Deech08
I was originally planning to fix this with a file check / redownload but I think I will just try and include it in the package instead. I believe I’ll need to request a file size limit increase on PyPi so I will update both issues when it’s done!

Thanks for catching these errors - I haven’t necessarily been able to reproduce them but I think I can if I interrupt the download.

> On Dec 12, 2019, at 12:10, Heloise <notifications@github.com> wrote:
>
> I had the same problem and then the following issue occured
>
> It would be worth warning the user that this is a big file and will take a while to download if it's not added to the package.
>
