### 2019-12-12 4179ab3
Upgrade to v0.2.0a1
Add survey data to package distributiion by default
