+++ MANIFEST.in
@@ -4,10 +4,10 @@ include CHANGES.rst
 include ah_bootstrap.py
 include setup.cfg
 include whampy/data/Reid16_SpiralArms/*
+include whampy/data/wham-ss-DR1-v161116-170912.fits
 
 exclude whampy/data/wham-ss-161116.sav
 exclude whampy/data/sii_161221.sav
-exclude whampy/data/wham-ss-DR1-v161116-170912.fits
 exclude whampy/data/wham-ss-DR1-v161116-170912.sav
 
 
