### 2019-12-12 Deech08/whampy#2 @Deech08
Latest version (v0.2.0a1) is now pip installable and updated on the repo. This release includes the survey data in the package itself and no additional download is necessary. So as long as the package installs it should be good to go.
