### 2019-12-12 Deech08/whampy#3 @Deech08
See fix to #2
