### 2019-12-12 fc41633
Update Readme
Fix Broken WHAM link and add clear link to Read The Docs
