+++ setup.cfg
@@ -41,7 +41,7 @@ github_project = astropy/astropy
 # install_requires = astropy, scipy, matplotlib, cartopy, spectral_cube
 install_requires = astropy, numpy, matplotlib, scipy, seaborn
 # version should be PEP440 compatible (https://www.python.org/dev/peps/pep-0440/)
-version = 0.2.0a1
+version = 0.2.0a2
 # Note: you will also need to change this in your package's __init__.py
 minimum_python_version = 3.6
 
