### 2019-12-12 Deech08/whampy#4 @Deech08
I've added a section after the Quick Start that links explicitly to the documentation. I was hoping users would just click on the doc:passing badge, but see that isn't super obvious.

The broken WHAM link is also now fixed - it directs to a webpage on the instrument itself.
