### 2019-12-13 Deech08/whampy#5 @HeloiseS
RuntimeWarning in QuickStart instructions
## Issue
Tried the quick start instructions after downloadig v0.2.0a1 and I get the following:
```
>>> moment_0, err_0 = survey.moment(order = 0, return_sigma = True)
/home/fste075/anaconda3/envs/whampy_test/lib/python3.7/site-packages/astropy/table/column.py:991: RuntimeWarning: invalid value encountered in less result = getattr(super(), op)(other)
```

The other two moments and errors calculate fine with no error :)

-----
*This issue is part of the current JOSS-review of whampy ( [#1940](https://github.com/openjournals/joss-reviews/issues/1940) )*
