### 2019-12-13 Deech08/whampy#6 @HeloiseS
Community Guidelines 
One of the requirements for JOSS publication is the presence of clearly stated Community Guidelines

> *Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support*

This could be added to a section in the README - feel free to inspire yourself from the astropy stuff :) This should be a quick fix

-----
*This issue is part of the current JOSS-review of whampy ( [#1940](https://github.com/openjournals/joss-reviews/issues/1940) )*
