### 2019-12-13 Deech08/whampy#5 @Deech08
I believe this happens because there are some nan values in the data. It should still compute fine, but just display the warning. Does the end result still give you actual numbers for moment_0 and err_0, or is there an actual error beyond the warning that stops it from executing?
