### 2019-12-14 cefd59d
Update issue templates
Add a bug reporting issue template
