### 2019-12-14 Deech08/whampy#6 @Deech08
I am planning to add a section to the README like the following. I believe this will satisfy those requirements. I previously had some brief text on the repo's ["Wiki" page](https://github.com/Deech08/whampy/wiki) but see that it isn't very clearly located. I added a couple issue templates to assist with the reporting process as well.

Contributing
------------

Contributions, feedback, and bug reports are welcome from all users of whampy! If you discover a bug, or have a request for a new feature, please open an issue on this repository using one of the templates. If you would like to make a more direct contribution to the code, please submit a pull request to this repository. This package relies upon the work of the astropy team, and we also ask that all contributions follow the `Astropy Community Code of Conduct <https://www.astropy.org/about.html#codeofconduct>`_.
