### 2019-12-15 #1940 @xuanxu
@HeloiseS: to have an explicit list of requierements is not a mandatory or stopper condition to pass a JOSS review. But if you thinks it's a standard/good practice or you feel it is something a user would expect to find, it is totally ok to suggest the author to add that list to the readme (or add a requirements.txt file or whatever other solution you have in mind).
