### 2019-12-16 Deech08/whampy#5 @HeloiseS
Ace, are these the values I should get?

```
>>> moment_0
<Quantity [0.94620246, 1.1443384 , 1.2161877 , ..., 6.932495  , 7.1546907 ,
           7.571975  ] R>
>>> err_0
<Quantity [0.23768413, 0.25891924, 0.27316484, ..., 0.3584447 , 0.2074711 ,
           0.38070837] R>
```
