### 2019-12-16 Deech08/whampy#9 @HeloiseS
Remote loading of Sky Survey crashes - maybe obsolete? [BUG]
## Issue

I am following [this page](https://whampy.readthedocs.io/en/latest/loading_data.html) and get the following error.

Since you're now including the survey in the package, are you going to keep the remote loading feature? If yes, this needs looking out. Otherwise the documentation will need updating

```
>>> from whampy.skySurvey import SkySurvey
>>> survey_remote = SkySurvey(mode = 'remote')
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/fste075/anaconda3/envs/whampy_test/lib/python3.7/site-packages/whampy/skySurvey.py", line 239, in __init__
    with fits.open(filename) as hdulist:
  File "/home/fste075/anaconda3/envs/whampy_test/lib/python3.7/site-packages/astropy/io/fits/hdu/hdulist.py", line 151, in fitsopen
    lazy_load_hdus, **kwargs)
  File "/home/fste075/anaconda3/envs/whampy_test/lib/python3.7/site-packages/astropy/io/fits/hdu/hdulist.py", line 390, in fromfile
    lazy_load_hdus=lazy_load_hdus, **kwargs)
  File "/home/fste075/anaconda3/envs/whampy_test/lib/python3.7/site-packages/astropy/io/fits/hdu/hdulist.py", line 1039, in _readfrom
    fileobj = _File(fileobj, mode=mode, memmap=memmap, cache=cache)
  File "/home/fste075/anaconda3/envs/whampy_test/lib/python3.7/site-packages/astropy/utils/decorators.py", line 521, in wrapper
    return function(*args, **kwargs)
  File "/home/fste075/anaconda3/envs/whampy_test/lib/python3.7/site-packages/astropy/io/fits/file.py", line 147, in __init__
    self.name = download_file(fileobj, cache=cache)
  File "/home/fste075/anaconda3/envs/whampy_test/lib/python3.7/site-packages/astropy/utils/data.py", line 1012, in download_file
    with shelve.open(urlmapfn) as url2hash:
  File "/home/fste075/anaconda3/envs/whampy_test/lib/python3.7/shelve.py", line 243, in open
    return DbfilenameShelf(filename, flag, protocol, writeback)
  File "/home/fste075/anaconda3/envs/whampy_test/lib/python3.7/shelve.py", line 227, in __init__
    Shelf.__init__(self, dbm.open(filename, flag), protocol, writeback)
  File "/home/fste075/anaconda3/envs/whampy_test/lib/python3.7/dbm/__init__.py", line 91, in open
    "available".format(result))
dbm.error: db type is dbm.gnu, but the module is not available
```

## My Set up
I'm working in a clean python 3.7 anaconda env where I'd only pip installed whampy.

-----
*This issue is part of the current JOSS-review of whampy ( [#1940](https://github.com/openjournals/joss-reviews/issues/1940) )*
