### 2019-12-16 Deech08/whampy#5 @Deech08
Yes, those are the expected results!
