### 2019-12-17 Deech08/whampy#10 @HeloiseS
Making Moment Maps -- Missing `import` statments
## Issue

There are 2 missing imports in your tutorial [this tutorial ](https://whampy.readthedocs.io/en/latest/moment_maps.html) and then it works spot on :+1: . Just add:

```
import matplotlib.pyplot as plt
import cartopy.crs as ccrs
```

## Original Issue for history
>  I am testing [this tutorial ](https://whampy.readthedocs.io/en/latest/moment_maps.html) and received the following error.
>
> A bit of googling seems to reveal this is a cartopy dependency - I didn't install anything beyond whampy and the tutorial doesn't call for a cartopy import, so this could be it.
>
> ```
> >>> mean_vel = survey.moment(order=1)
> /home/fste075/anaconda3/envs/whampy_test/lib/python3.7/site-packages/astropy/table/column.py:991: RuntimeWarning: invalid value encountered in less
>   result = getattr(super(), op)(other)
> >>> fig = plt.figure()
> >>> ax = fig.add_subplot(111, projection = ccrs.Mollweide(central_longitude=0))
> Traceback (most recent call last):
>   File "<stdin>", line 1, in <module>
> NameError: name 'ccrs' is not defined
>```

-----
*This issue is part of the current JOSS-review of whampy ( [#1940](https://github.com/openjournals/joss-reviews/issues/1940) )*
