### 2019-12-17 Deech08/whampy#11 @HeloiseS
Dependency List Needed [FEATURE]
I noticed the [Making Interactive Maps with Whampy tutorial ](https://whampy.readthedocs.io/en/latest/click_maps.html) also requires cartopy, this time explicitly. Since it seems important to core features of the utilities you produced it should either be added to the dependencies in the pip install or explicitly stated in the READ ME.

(also your interactive map is absolute bomb - really awesome :smiley: )

-----
*This issue is part of the current JOSS-review of whampy ( [#1940](https://github.com/openjournals/joss-reviews/issues/1940) )*
