### 2019-12-17 Deech08/whampy#12 @HeloiseS
spectral_cube dependency not installed with pip [BUG]
## Issue

I was playing with  the last part of the [Interactive Map tutorial](https://whampy.readthedocs.io/en/latest/click_maps.html) and got this error:
```
>>> click_map = survey.click_map(fig = fig, over_data = fits_cube_path)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/fste075/anaconda3/envs/whampy_test/lib/python3.7/site-packages/whampy/whampyTableMixin.py", line 499, in click_map
    radius = radius)
  File "/home/fste075/anaconda3/envs/whampy_test/lib/python3.7/site-packages/whampy/clickMap.py", line 78, in __init__
    self.over_data = SpectralCube.read(over_data)
NameError: name 'SpectralCube' is not defined
```

I got it running by installing spectral cube separately but since whampy is pip installed this dependency can be taken care of on download.

-----
*This issue is part of the current JOSS-review of whampy ( [#1940](https://github.com/openjournals/joss-reviews/issues/1940) )*
