### 2019-12-17 Deech08/whampy#13 @HeloiseS
Spectral cube handling fails[BUG]
## Issue

So I tried parsing a data cube I had lying around to see what would happen at the end of [this tutorial](https://whampy.readthedocs.io/en/latest/click_maps.html).

I got the following error.

```
>>> click_map = survey.click_map(fig = fig, over_data = fits_cube_path)
WARNING: FITSWarning: hdu= was not specified but multiple arrays are present, reading in first available array (hdu=0) [spectral_cube.io.fits]
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/fste075/anaconda3/envs/whampy_test/lib/python3.7/site-packages/whampy/whampyTableMixin.py", line 499, in click_map
    radius = radius)
  File "/home/fste075/anaconda3/envs/whampy_test/lib/python3.7/site-packages/whampy/clickMap.py", line 78, in __init__
    self.over_data = SpectralCube.read(over_data)
  File "/home/fste075/anaconda3/envs/whampy_test/lib/python3.7/site-packages/spectral_cube/spectral_cube.py", line 2172, in read
    cube = read(filename, format=format, hdu=hdu, **kwargs)
  File "/home/fste075/anaconda3/envs/whampy_test/lib/python3.7/site-packages/spectral_cube/io/core.py", line 33, in read
    return load_fits_cube(filename, hdu=hdu, **kwargs)
  File "/home/fste075/anaconda3/envs/whampy_test/lib/python3.7/site-packages/spectral_cube/io/fits.py", line 142, in load_fits_cube
    'keyword argument to read data from another HDU.'.format(hdu))
spectral_cube.utils.FITSReadError: No data found in HDU None. You can try using the hdu= keyword argument to read data from another HDU.
>>>
```

## Note

I don't necessarily expect you to make this versatile enough to accommodate all data cubes, but telling users in the tutorial what kind of data they will be able to parse would be useful.

-----
*This issue is part of the current JOSS-review of whampy ( [#1940](https://github.com/openjournals/joss-reviews/issues/1940) )*
