### 2019-12-17 Deech08/whampy#9 @Deech08
I agree that its probably just best to remove the remote feature all together. I will update docs and close the issue when that is done.
