### 2019-12-17 Deech08/whampy#11 @Deech08
I'm glad you like the interactive map!

As for cartopy - I originally had it as a dependency and removed it because it sometimes causes install issues (especially on linux machines apparently). I was favoring making it an optional dependency. Would it suffice to add more specific statements mentioning on that tutorial that this feature will only work if you install an optional other package (cartopy)?
