### 2019-12-17 Deech08/whampy#13 @Deech08
I think this would be a spectral-cube issue. Do you mind sharing some more details of the data cube you tried parsing? maybe just what the header looks like?
