### 2019-12-17 Deech08/whampy#12 @Deech08
I am modifying this tutorial to mention that the cartopy and spectral cube examples rely on those packages which are optional dependencies. I hope that this would prompt the user to manually install those packages as well if they want to run that part of the tutorial. I would prefer to leave both packages as optional dependencies, but can incorporate them in more directly if that is necessary.
