### 2019-12-17 d9a3f0b
Documentation Fixes
Corrent typos and documenation issues. Fix typo in moment computing
