+++ whampy/whampyTableMixin.py
@@ -193,12 +193,12 @@ class SkySurveyMixin(object):
                 moment_2 = np.trapz(data_masked * (vel_masked - moment_1.value[:,None])**2, 
                     x = vel_masked, 
                     axis = 1) * self["DATA"].unit * self["VELOCITY"].unit**3 / moment_0
-                print(moment_2.unit)
+
                 err_2_subover_mom2 = np.trapz(data_masked * (vel_masked - moment_1.value[:,None])**2 * np.sqrt(var_masked / 
                                 data_masked**2 + 2*(err_1[:,None] / moment_1[:,None])**2), 
                     x = vel_masked, 
                     axis = 1) * self["DATA"].unit * self["VELOCITY"].unit**3 / (moment_2 * moment_0)
-                print(err_2_subover_mom2.unit)
+
                 err_2 = moment_2 * np.sqrt(err_2_subover_mom2**2 + (err_0 / moment_0)**2)
                 if return_sigma:
                     return moment_2, err_2
