### 2019-12-17 Deech08/whampy#12 @HeloiseS
Sound fine to me! Would you mind making it very clear and explicit at the beginning of the tutorial that these need to be installed by the user? I know this may seem like overkill but I remember being new to python and people stating what was obvious to them would have tremendously helped me at times ;)
