### 2019-12-19 Deech08/whampy#11 @HeloiseS
I totally get that - I did have to fiddle a little to install cartopy. I think it is important to make it super clear that this is optional and something that the user needs to do. Maybe even sign post them to resources to install cartopy and its dependencies. This may feel like over kill but I'm trying to keep in mind junior scientists or people who are new to python and don't have the same reflexes we do in terms of debugging. :)
