### 2019-12-19 Deech08/whampy#10 @HeloiseS
Is it a lot of hassle to make a new pip version when you've fixed most of the initial issues we flagged?

This allows me to create a completely fresh environment in conda and pip install it anew as if I was a first time user :) I find this is the a great way to shoo the bugs away.
