### 2019-12-19 Deech08/whampy#10 @HeloiseS
Also, I should have said that the last two commands seem to have an extra indent:

```
>>> fig = plt.figure()
    >>> ax = fig.add_subplot(111, projection = ccrs.Mollweide(central_longitude=0))

    >>> fig = survey.intensity_map(ax = ax, c = mean_vel, vmin = -75, vmax = 75,
                                       norm = None, cmap = "RdBu_r", colorbar = True,
                                       cbar_kwargs={"orientation":"horizontal",
                                                            "label":"Mean Velocity (km/s)"})
```
