### 2019-12-19 Deech08/whampy#13 @HeloiseS
The header is a beast of a MUSE header - sorry :fearful:

```
SIMPLE  =                    T / file does conform to FITS standard
BITPIX  =                    8 / number of bits per data pixel
NAXIS   =                    0 / number of data axes
EXTEND  =                    T / FITS dataset may contain extensions
COMMENT   FITS (Flexible Image Transport System) format is defined in 'Astronomy
COMMENT   and Astrophysics', volume 376, page 359; bibcode: 2001A&A...376..359H
DATE    = '2017-09-22T14:02:14' / file creation date (YYYY-MM-DDThh:mm:ss UT)
ORIGIN  = 'ESO-PARANAL'        / European Southern Observatory
TELESCOP= 'ESO-VLT-U4'         / ESO <TEL>
INSTRUME= 'MUSE    '           / Instrument used.
RA      =     197.451801028829 / [deg] Image center (J2000)
DEC     =    -23.3798559057508 / [deg] Image center (J2000)
EQUINOX =                2000. / Standard FK5
RADECSYS= 'FK5     '           / Coordinate system
EXPTIME =      1809.4369094199 / [s] Total integration time per pixel
TEXPTIME=                2600. / [s] Total integration time of all exposures
NCOMBINE=                    4 / No. of combined raw science data files
MJD-OBS =       57983.99298902 / [d] Start of observations (days)
MJD-END =     57984.0270303581 / [d] End of observations (days)
DATE-OBS= '2017-08-18T23:49:54.000' / Observing date
UTC     =               85782. / [s] 23:49:42.000 UTC
LST     =            61091.982 / [s] 16:58:11.982 LST
PI-COI  = 'UNKNOWN '           / PI-COI name.
OBSERVER= 'UNKNOWN '           / Name of observer.
PIPEFILE= 'r.MUSE.2017-08-18T23:49:54.251_tpl_0000.fits' / no comment
BUNIT   = '10**(-20)*erg/s/cm**2/Angstrom'
ARCFILE = 'ADP.2017-09-22T14:07:52.337.fits' / Archive File Name
DATAMD5 = 'Not computed'       / MD5 checksum
OBJECT  = 'SGRB    '           / Original target.
OBID1   =              1690060 / Observation block ID
PROG_ID = '099.D-0668(B)'      / ESO programme identification
PROV1   = 'MUSE.2017-08-18T23:49:54.251.fits'
PROV2   = 'MUSE.2017-08-19T00:02:37.713.fits'
PROV3   = 'MUSE.2017-08-19T00:15:21.698.fits'
PROV4   = 'MUSE.2017-08-19T00:28:05.423.fits'
ASSON1  = 'MU_SIMC_1690060_2017-08-18T23:49:54.251_WFM-NOAO-N_OBJ.fits'
PRODCATG= 'SCIENCE.CUBE.IFS'   / Data product category
PROCSOFT= 'muse/2.0.3'         / ESO pipeline version
OBSTECH = 'IFU     '           / Technique for observation
FLUXCAL = 'ABSOLUTE'           / Type of flux calibration (ABSOLUTE or UNCALIBRA
WAVELMIN=     474.961767578125 / [nm] Minimum wavelength
WAVELMAX=     934.961767578125 / [nm] Maximum wavelength
SPEC_RES=     3026.52327095622 / Spectral resolving power at central wavelength
SKY_RES =     1.44324018548212 / [arcsec] FWHM effective spatial resolution (mea
SKY_RERR=    0.309232921921325 / [arcsec] Error of SKY_RES (estimated)
PIXNOISE=  4.9457052075513E-20 / [erg/s/cm**2/Angstrom] pixel-to-pixel noise
ABMAGLIM=     21.5131147116938 / 5-sigma magnitude limit for point sources
REFERENC= '        '           / Reference publication
HIERARCH ESO OBS AIRM =    2.8 / Req. max. airmass
HIERARCH ESO OBS AMBI FWHM = 1.6 / Req. max. seeing
HIERARCH ESO OBS AMBI TRANS = '3THN    ' / Req. sky transparency
HIERARCH ESO OBS ATM = 'NATM    ' / Req. Atmospheric Turbulence Model
HIERARCH ESO OBS CONTAINER ID = -999 / Scheduling container ID
HIERARCH ESO OBS CONTAINER TYPE = 'U       ' / Scheduling container type
HIERARCH ESO OBS CONTRAST = 0. / Req. contrast
HIERARCH ESO OBS DID = 'ESO-VLT-DIC.OBS-2.0' / OBS Dictionary
HIERARCH ESO OBS EXECTIME = 3491 / Expected execution time
HIERARCH ESO OBS GRP = '0       ' / linked blocks
HIERARCH ESO OBS ID =  1690060 / Observation block ID
HIERARCH ESO OBS MOON DIST = 30 / Req. min. angular dist. from moon [deg]
HIERARCH ESO OBS MOON FLI = 1. / Req. max. fractional lunar illum.
HIERARCH ESO OBS NAME = 'TOO_WFM_MUSE_SGRB' / OB name
HIERARCH ESO OBS NTPL =      2 / Number of templates within OB
HIERARCH ESO OBS OBSERVER = 'UNKNOWN ' / Observer Name
HIERARCH ESO OBS PI-COI ID = 13037 / ESO internal PI-COI ID
HIERARCH ESO OBS PI-COI NAME = 'UNKNOWN ' / PI-COI name
HIERARCH ESO OBS PROG ID = '099.D-0668(B)' / ESO program identification
HIERARCH ESO OBS START = '2017-08-18T23:46:59' / OB start time
HIERARCH ESO OBS STREHLRATIO = 0. / Req. strehl ratio
HIERARCH ESO OBS TARG NAME = 'SGRB    ' / OB target name
HIERARCH ESO OBS TPLNO =     2 / Template number within OB
HIERARCH ESO OBS TWILIGHT =  0 / Req. twilight
HIERARCH ESO OBS WATERVAPOUR = 0. / Req. water vapour
HIERARCH ESO TPL DID = 'ESO-VLT-DIC.TPL-1.9' / Data dictionary for TPL
HIERARCH ESO TPL EXPNO =     1 / Exposure number within template
HIERARCH ESO TPL ID = 'MUSE_wfm-noao_obs_genericoffset' / Template signature ID
HIERARCH ESO TPL NAME = 'Observation in WFM with user defined offsets' / Templat
HIERARCH ESO TPL NEXP =      4 / Number of exposures within template
HIERARCH ESO TPL PRESEQ = 'MUSE_obs_offset.seq' / Sequencer script
HIERARCH ESO TPL START = '2017-08-18T23:49:38' / TPL start time
HIERARCH ESO TPL VERSION = '$Revision: 297587 $' / Version of the template
HIERARCH ESO TEL AIRM END = 1.707 / Airmass at end
HIERARCH ESO TEL AIRM START = 1.606 / Airmass at start
HIERARCH ESO TEL ALT =  38.454 / [deg] Alt angle at start
HIERARCH ESO TEL AMBI FWHM END = 0.74 / [arcsec] Observatory Seeing queried from
HIERARCH ESO TEL AMBI FWHM START = 0.72 / [arcsec] Observatory Seeing queried fr
HIERARCH ESO TEL AMBI IRSKY TEMP = -98.1 / Temperature of the IR sky, from radio
HIERARCH ESO TEL AMBI IWV END = 1.23 / Integrated Water Vapor
HIERARCH ESO TEL AMBI IWV START = 1.23 / Integrated Water Vapor
HIERARCH ESO TEL AMBI IWV30D END = 1.33 / IWV at 30deg elev.
HIERARCH ESO TEL AMBI IWV30D START = 1.34 / IWV at 30deg elev.
HIERARCH ESO TEL AMBI IWV30DSTD END = 0.02 / IWV at 30deg elev.
HIERARCH ESO TEL AMBI IWV30DSTD START = 0.01 / IWV at 30deg elev.
HIERARCH ESO TEL AMBI IWVSTD END = 0.01 / Standard Deviation of Integrated Water
HIERARCH ESO TEL AMBI IWVSTD START = 0.01 / Standard Deviation of Integrated Wat
HIERARCH ESO TEL AMBI PRES END = 744.57 / [hPa] Observatory ambient air pressure
HIERARCH ESO TEL AMBI PRES START = 744.48 / [hPa] Observatory ambient air pressu
HIERARCH ESO TEL AMBI RHUM = 10.5 / [%] Observatory ambient relative humidity qu
HIERARCH ESO TEL AMBI TAU0 = 0.0035 / [s] Average coherence time
HIERARCH ESO TEL AMBI TEMP = 10.93 / [deg C] Observatory ambient temperature que
HIERARCH ESO TEL AMBI WINDDIR = 53.5 / [deg] Observatory ambient wind direction
HIERARCH ESO TEL AMBI WINDSP = 2.75 / [m/s] Observatory ambient wind speed queri
HIERARCH ESO TEL AZ =   78.723 / [deg] Az angle at start S=0,W=90
HIERARCH ESO TEL CHOP ST =   F / True when chopping is active
HIERARCH ESO TEL DATE = '2000-01-01T00:00:00' / TCS installation date
HIERARCH ESO TEL DID = 'ESO-VLT-DIC.TCS' / Data dictionary for TEL
HIERARCH ESO TEL DOME STATUS = 'FULLY-OPEN' / Dome status
HIERARCH ESO TEL FO POS =   0. / Adapter focus position during tracking
HIERARCH ESO TEL FOCU ID = 'NB      ' / Telescope focus station ID
HIERARCH ESO TEL FOCU LEN = 121.561 / [m] Focal length
HIERARCH ESO TEL FOCU SCALE = 1.705 / [arcsec/mm] Focal scale
HIERARCH ESO TEL FOCU VALUE = -19.35 / [mm] M2 setting
HIERARCH ESO TEL GEOELEV = 2648. / [m] Elevation above sea level
HIERARCH ESO TEL GEOLAT = -24.627 / [deg] Tel geo latitute (+=North)
HIERARCH ESO TEL GEOLON = -70.404 / [deg] Tel geo longitude (+=East)
HIERARCH ESO TEL IA FWHM = 0.88 / [arcsec] Delivered seeing corrected by airmass
HIERARCH ESO TEL IA FWHMLIN = 1.28 / Delivered seeing on IA detector (linear fit
HIERARCH ESO TEL IA FWHMLINOBS = 1.76 / Delivered seeing on IA detector (linear
HIERARCH ESO TEL ID = 'v 299841' / TCS version number
HIERARCH ESO TEL MOON DEC = 19.20797 / [deg] 19:12:28.6 DEC (J2000)
HIERARCH ESO TEL MOON RA = 109.649633 / [deg] 07:18:35.9 RA (J2000)
HIERARCH ESO TEL OPER = 'I, Condor' / Telescope Operator
HIERARCH ESO TEL PARANG END = 104.3 / [deg] Parallactic angle at end
HIERARCH ESO TEL PARANG START = 103.607 / [deg] Parallactic angle at start
HIERARCH ESO TEL ROT ALTAZTRACK = T / Track rotator on alt/az instead of Alpha/D
HIERARCH ESO TEL TARG ALPHA = 130948.08 / Alpha coordinate for the target
HIERARCH ESO TEL TARG COORDTYPE = 'M       ' / Coordinate type (M=mean A=apparen
HIERARCH ESO TEL TARG DELTA = -232252.2 / Delta coordinate for the target
HIERARCH ESO TEL TARG EPOCH = 2000. / [yr] Epoch 1950 or 2000
HIERARCH ESO TEL TARG EPOCHSYSTEM = 'J       ' / Epoch system (J=Julian / B=Bess
HIERARCH ESO TEL TARG EQUINOX = 2000. / [yr] Equinox
HIERARCH ESO TEL TARG PARALLAX = 0. / [arcsec] Parallax
HIERARCH ESO TEL TARG PMA = 0. / [arcsec/yr] Proper Motion Alpha
HIERARCH ESO TEL TARG PMD = 0. / [arcsec/yr] Proper motion Delta
HIERARCH ESO TEL TARG RADVEL = 0. / [km/s] Radial velocity
HIERARCH ESO TEL TH M1 TEMP = 10.03 / [deg C] M1 superficial temperature
HIERARCH ESO TEL TRAK STATUS = 'NORMAL  ' / Tracking status
HIERARCH ESO INS ADC1 DEC = -232823.15146 / [deg] Telescope declination
HIERARCH ESO INS ADC1 ENC END = 190673 / [Enc] ADC encoder position at end
HIERARCH ESO INS ADC1 ENC START = 190673 / [Enc] ADC encoder position at start
HIERARCH ESO INS ADC1 END = -89.9999 / [deg] ADC position angle at end
HIERARCH ESO INS ADC1 MODE = 'OFF     ' / ADC mode
HIERARCH ESO INS ADC1 RA = 131044.119668 / [deg] Telescope right ascension
HIERARCH ESO INS ADC2 DEC = -232823.15146 / [deg] Telescope declination
HIERARCH ESO INS ADC2 ENC END = 230673 / [Enc] ADC encoder position at end
HIERARCH ESO INS ADC2 ENC START = 230673 / [Enc] ADC encoder position at start
HIERARCH ESO INS ADC2 END = -270. / [deg] ADC position angle at end
HIERARCH ESO INS ADC2 MODE = 'OFF     ' / ADC mode
HIERARCH ESO INS ADC2 RA = 131044.119668 / [deg] Telescope right ascension
HIERARCH ESO INS AMPL1 ID = 'ampl1   ' / Pico-amplifier ID
HIERARCH ESO INS AMPL1 SWSIM = F / If T, pico-amplifier software simulated
HIERARCH ESO INS AMPL2 FILTER = 'Kron_V  ' / Name of filter in front of PD2
HIERARCH ESO INS AMPL2 ID = 'ampl2   ' / Pico-amplifier ID
HIERARCH ESO INS AMPL2 SWSIM = F / If T, pico-amplifier software simulated
HIERARCH ESO INS DATE = '2000-06-16' / Instrument release date (yyyy-mm-dd).
HIERARCH ESO INS DID1 = 'MUSE_ICS-297470' / Data dictionary for INSi
HIERARCH ESO INS DID2 = 'PRIMARY-FITS-1.2' / Data dictionary for INSi
HIERARCH ESO INS DROT DEC = -232823.15146 / [deg] Telescope declination
HIERARCH ESO INS DROT ENC END = 109122668 / [Enc] Derotator encoder position at
HIERARCH ESO INS DROT ENC START = 108781470 / [Enc] Derotator encoder position a
HIERARCH ESO INS DROT END = -70.0373 / [deg] Derotator position angle at end
HIERARCH ESO INS DROT MODE = 'SKY     ' / Instrument derotator mode
HIERARCH ESO INS DROT POSANG = 0. / [deg] Derotator position angle
HIERARCH ESO INS DROT RA = 131044.119668 / [deg] Telescope right ascension
HIERARCH ESO INS ID = 'MUSE/176642' / Instrument ID
HIERARCH ESO INS IFU1 CABLE ID = '-1      ' / Detector cable ID
HIERARCH ESO INS IFU1 CHAN = 2 / Integral Field Unit channel
HIERARCH ESO INS IFU1 DV ID = '16      ' / Detector Vessel ID
HIERARCH ESO INS IFU1 ID = '01      ' / Integral Field Unit ID
HIERARCH ESO INS IFU1 ISS ID = '24      ' / Image Slicer System ID
HIERARCH ESO INS IFU1 PREAMP ID = '36      ' / Detector cable ID
HIERARCH ESO INS IFU1 SPS ID = '01      ' / Spectograph System ID
HIERARCH ESO INS IFU1 VPHG ID = '08      ' / Volume Phase Holographic Grating ID
HIERARCH ESO INS IFU10 CABLE ID = '-1      ' / Detector cable ID
HIERARCH ESO INS IFU10 CHAN = 3 / Integral Field Unit channel
HIERARCH ESO INS IFU10 DV ID = '25      ' / Detector Vessel ID
HIERARCH ESO INS IFU10 ID = '10      ' / Integral Field Unit ID
HIERARCH ESO INS IFU10 ISS ID = '10      ' / Image Slicer System ID
HIERARCH ESO INS IFU10 PREAMP ID = '29      ' / Detector cable ID
HIERARCH ESO INS IFU10 SPS ID = '10      ' / Spectograph System ID
HIERARCH ESO INS IFU10 VPHG ID = '19      ' / Volume Phase Holographic Grating I
HIERARCH ESO INS IFU11 CABLE ID = '-1      ' / Detector cable ID
HIERARCH ESO INS IFU11 CHAN = 12 / Integral Field Unit channel
HIERARCH ESO INS IFU11 DV ID = '10      ' / Detector Vessel ID
HIERARCH ESO INS IFU11 ID = '11      ' / Integral Field Unit ID
HIERARCH ESO INS IFU11 ISS ID = '03      ' / Image Slicer System ID
HIERARCH ESO INS IFU11 PREAMP ID = '55      ' / Detector cable ID
HIERARCH ESO INS IFU11 SPS ID = '11      ' / Spectograph System ID
HIERARCH ESO INS IFU11 VPHG ID = '15      ' / Volume Phase Holographic Grating I
HIERARCH ESO INS IFU12 CABLE ID = '-1      ' / Detector cable ID
HIERARCH ESO INS IFU12 CHAN = 7 / Integral Field Unit channel
HIERARCH ESO INS IFU12 DV ID = '21      ' / Detector Vessel ID
HIERARCH ESO INS IFU12 ID = '12      ' / Integral Field Unit ID
HIERARCH ESO INS IFU12 ISS ID = '09      ' / Image Slicer System ID
HIERARCH ESO INS IFU12 PREAMP ID = '50      ' / Detector cable ID
HIERARCH ESO INS IFU12 SPS ID = '12      ' / Spectograph System ID
HIERARCH ESO INS IFU12 VPHG ID = '21      ' / Volume Phase Holographic Grating I
HIERARCH ESO INS IFU13 CABLE ID = '-1      ' / Detector cable ID
HIERARCH ESO INS IFU13 CHAN = 21 / Integral Field Unit channel
HIERARCH ESO INS IFU13 DV ID = '22      ' / Detector Vessel ID
HIERARCH ESO INS IFU13 ID = '13      ' / Integral Field Unit ID
HIERARCH ESO INS IFU13 ISS ID = '08      ' / Image Slicer System ID
HIERARCH ESO INS IFU13 PREAMP ID = '78      ' / Detector cable ID
HIERARCH ESO INS IFU13 SPS ID = '13      ' / Spectograph System ID
HIERARCH ESO INS IFU13 VPHG ID = '18      ' / Volume Phase Holographic Grating I
HIERARCH ESO INS IFU14 CABLE ID = '-1      ' / Detector cable ID
HIERARCH ESO INS IFU14 CHAN = 5 / Integral Field Unit channel
HIERARCH ESO INS IFU14 DV ID = '23      ' / Detector Vessel ID
HIERARCH ESO INS IFU14 ID = '14      ' / Integral Field Unit ID
HIERARCH ESO INS IFU14 ISS ID = '12      ' / Image Slicer System ID
HIERARCH ESO INS IFU14 PREAMP ID = '34      ' / Detector cable ID
HIERARCH ESO INS IFU14 SPS ID = '14      ' / Spectograph System ID
HIERARCH ESO INS IFU14 VPHG ID = '20      ' / Volume Phase Holographic Grating I
HIERARCH ESO INS IFU15 CABLE ID = '-1      ' / Detector cable ID
HIERARCH ESO INS IFU15 CHAN = 15 / Integral Field Unit channel
HIERARCH ESO INS IFU15 DV ID = '26      ' / Detector Vessel ID
HIERARCH ESO INS IFU15 ID = '15      ' / Integral Field Unit ID
HIERARCH ESO INS IFU15 ISS ID = '15      ' / Image Slicer System ID
HIERARCH ESO INS IFU15 PREAMP ID = '66      ' / Detector cable ID
HIERARCH ESO INS IFU15 SPS ID = '15      ' / Spectograph System ID
HIERARCH ESO INS IFU15 VPHG ID = '22      ' / Volume Phase Holographic Grating I
HIERARCH ESO INS IFU16 CABLE ID = '-1      ' / Detector cable ID
HIERARCH ESO INS IFU16 CHAN = 18 / Integral Field Unit channel
HIERARCH ESO INS IFU16 DV ID = '27      ' / Detector Vessel ID
HIERARCH ESO INS IFU16 ID = '16      ' / Integral Field Unit ID
HIERARCH ESO INS IFU16 ISS ID = '16      ' / Image Slicer System ID
HIERARCH ESO INS IFU16 PREAMP ID = '57      ' / Detector cable ID
HIERARCH ESO INS IFU16 SPS ID = '16      ' / Spectograph System ID
HIERARCH ESO INS IFU16 VPHG ID = '17      ' / Volume Phase Holographic Grating I
HIERARCH ESO INS IFU17 CABLE ID = '-1      ' / Detector cable ID
HIERARCH ESO INS IFU17 CHAN = 13 / Integral Field Unit channel
HIERARCH ESO INS IFU17 DV ID = '17      ' / Detector Vessel ID
HIERARCH ESO INS IFU17 ID = '17      ' / Integral Field Unit ID
HIERARCH ESO INS IFU17 ISS ID = '13      ' / Image Slicer System ID
HIERARCH ESO INS IFU17 PREAMP ID = '46      ' / Detector cable ID
HIERARCH ESO INS IFU17 SPS ID = '17      ' / Spectograph System ID
HIERARCH ESO INS IFU17 VPHG ID = '16      ' / Volume Phase Holographic Grating I
HIERARCH ESO INS IFU18 CABLE ID = '-1      ' / Detector cable ID
HIERARCH ESO INS IFU18 CHAN = 9 / Integral Field Unit channel
HIERARCH ESO INS IFU18 DV ID = '08      ' / Detector Vessel ID
HIERARCH ESO INS IFU18 ID = '18      ' / Integral Field Unit ID
HIERARCH ESO INS IFU18 ISS ID = '19      ' / Image Slicer System ID
HIERARCH ESO INS IFU18 PREAMP ID = '43      ' / Detector cable ID
HIERARCH ESO INS IFU18 SPS ID = '18      ' / Spectograph System ID
HIERARCH ESO INS IFU18 VPHG ID = '26      ' / Volume Phase Holographic Grating I
HIERARCH ESO INS IFU19 CABLE ID = '-1      ' / Detector cable ID
HIERARCH ESO INS IFU19 CHAN = 14 / Integral Field Unit channel
HIERARCH ESO INS IFU19 DV ID = '13      ' / Detector Vessel ID
HIERARCH ESO INS IFU19 ID = '19      ' / Integral Field Unit ID
HIERARCH ESO INS IFU19 ISS ID = '18      ' / Image Slicer System ID
HIERARCH ESO INS IFU19 PREAMP ID = '37      ' / Detector cable ID
HIERARCH ESO INS IFU19 SPS ID = '19      ' / Spectograph System ID
HIERARCH ESO INS IFU19 VPHG ID = '23      ' / Volume Phase Holographic Grating I
HIERARCH ESO INS IFU2 CABLE ID = '-1      ' / Detector cable ID
HIERARCH ESO INS IFU2 CHAN = 6 / Integral Field Unit channel
HIERARCH ESO INS IFU2 DV ID = '24      ' / Detector Vessel ID
HIERARCH ESO INS IFU2 ID = '02      ' / Integral Field Unit ID
HIERARCH ESO INS IFU2 ISS ID = '06      ' / Image Slicer System ID
HIERARCH ESO INS IFU2 PREAMP ID = '52      ' / Detector cable ID
HIERARCH ESO INS IFU2 SPS ID = '02      ' / Spectograph System ID
HIERARCH ESO INS IFU2 VPHG ID = '03      ' / Volume Phase Holographic Grating ID
HIERARCH ESO INS IFU20 CABLE ID = '-1      ' / Detector cable ID
HIERARCH ESO INS IFU20 CHAN = 10 / Integral Field Unit channel
HIERARCH ESO INS IFU20 DV ID = '20      ' / Detector Vessel ID
HIERARCH ESO INS IFU20 ID = '20      ' / Integral Field Unit ID
HIERARCH ESO INS IFU20 ISS ID = '20      ' / Image Slicer System ID
HIERARCH ESO INS IFU20 PREAMP ID = '38      ' / Detector cable ID
HIERARCH ESO INS IFU20 SPS ID = '20      ' / Spectograph System ID
HIERARCH ESO INS IFU20 VPHG ID = '24      ' / Volume Phase Holographic Grating I
HIERARCH ESO INS IFU21 CABLE ID = '-1      ' / Detector cable ID
HIERARCH ESO INS IFU21 CHAN = 11 / Integral Field Unit channel
HIERARCH ESO INS IFU21 DV ID = '15      ' / Detector Vessel ID
HIERARCH ESO INS IFU21 ID = '21      ' / Integral Field Unit ID
HIERARCH ESO INS IFU21 ISS ID = '17      ' / Image Slicer System ID
HIERARCH ESO INS IFU21 PREAMP ID = '23      ' / Detector cable ID
HIERARCH ESO INS IFU21 SPS ID = '21      ' / Spectograph System ID
HIERARCH ESO INS IFU21 VPHG ID = '27      ' / Volume Phase Holographic Grating I
HIERARCH ESO INS IFU22 CABLE ID = '-1      ' / Detector cable ID
HIERARCH ESO INS IFU22 CHAN = 24 / Integral Field Unit channel
HIERARCH ESO INS IFU22 DV ID = '14      ' / Detector Vessel ID
HIERARCH ESO INS IFU22 ID = '22      ' / Integral Field Unit ID
HIERARCH ESO INS IFU22 ISS ID = '22      ' / Image Slicer System ID
HIERARCH ESO INS IFU22 PREAMP ID = '54      ' / Detector cable ID
HIERARCH ESO INS IFU22 SPS ID = '22      ' / Spectograph System ID
HIERARCH ESO INS IFU22 VPHG ID = '11      ' / Volume Phase Holographic Grating I
HIERARCH ESO INS IFU23 CABLE ID = '-1      ' / Detector cable ID
HIERARCH ESO INS IFU23 CHAN = 23 / Integral Field Unit channel
HIERARCH ESO INS IFU23 DV ID = '18      ' / Detector Vessel ID
HIERARCH ESO INS IFU23 ID = '23      ' / Integral Field Unit ID
HIERARCH ESO INS IFU23 ISS ID = '23      ' / Image Slicer System ID
HIERARCH ESO INS IFU23 PREAMP ID = '60      ' / Detector cable ID
HIERARCH ESO INS IFU23 SPS ID = '23      ' / Spectograph System ID
HIERARCH ESO INS IFU23 VPHG ID = '31      ' / Volume Phase Holographic Grating I
HIERARCH ESO INS IFU24 CABLE ID = '-1      ' / Detector cable ID
HIERARCH ESO INS IFU24 CHAN = 1 / Integral Field Unit channel
HIERARCH ESO INS IFU24 DV ID = '12      ' / Detector Vessel ID
HIERARCH ESO INS IFU24 ID = '24      ' / Integral Field Unit ID
HIERARCH ESO INS IFU24 ISS ID = '02      ' / Image Slicer System ID
HIERARCH ESO INS IFU24 PREAMP ID = '24      ' / Detector cable ID
HIERARCH ESO INS IFU24 SPS ID = '24      ' / Spectograph System ID
HIERARCH ESO INS IFU24 VPHG ID = '10      ' / Volume Phase Holographic Grating I
[…]
