### 2019-12-19 #1940 @HeloiseS
Okay I have another question, this time about paper structure.

The author ticks pretty much all the boxes but everything is under only one section.
Is this an acceptable format for JOSS or would it be reasonable to ask for seperate sections to split the general summary from the specifics, or would that be redundant?

Sorry this is my first time reviewing for JOSS :smile:
