### 2019-12-19 Deech08/whampy#14 @HeloiseS
Astropy 2018 ref [PAPER]
I see you mention astropy in the JOSS paper, and you references the OG robitalle+2013 ref.
It would probably be good adding the astropy v2.0 reference as well: see [here](https://arxiv.org/abs/1801.02634)

-----
*This issue is part of the current JOSS-review of whampy ( [#1940](https://github.com/openjournals/joss-reviews/issues/1940) )*
