### 2019-12-19 #1940 @xuanxu
@HeloiseS only one section is acceptable, yes, JOSS papers are usually quite short ([some](https://joss.theoj.org/papers/10.21105/joss.01896) [examples](https://joss.theoj.org/papers/10.21105/joss.01881)). But in any case if you think that section is too long or it would be clearer with different sections, asking for a split is also ok. So I guess the answer is "both options are acceptable". :smiley:
