### 2019-12-19 Deech08/whampy#10 @Deech08
I will definitely make a new pip version once I make all these changes! I will probably do so in the next day or so and will post on these issues when it is done.
