### 2019-12-19 Deech08/whampy#13 @Deech08
Ah interesting - I don't believe spectral cube will work well with this kind of file (at least without some extra steps). I'll add some text explaining exactly what kinds of data can be used with spectral cube. Specifically, the data needs to be a 3d array of (position, position, velocity/frequency/wavelength) in a regular grid and the NAXIS keyword in the header needs to be 3.
