### 2019-12-19 f8d3261
Doc Updates
add to documentation and add astropy 2.0 paper reference
