### 2019-12-19 6e37a06
Update README.rst
Fix links for last two dependencies
