### 2019-12-19 Deech08/whampy#14 @Deech08
This citation should now be included in the lastest paper.md
