### 2019-12-19 Deech08/whampy#12 @Deech08
I've adding additional text in both the readme and docs mentioning they need to be installed separately and with links to their documentation.
