### 2019-12-19 Deech08/whampy#10 @Deech08
Fixed and incorporated into version 0.2.0a2
