### 2019-12-20 #1940 @HeloiseS
Hello all!

I think I'm pretty happy with the paper and the code as it stands now. Unless major changes come from @Gabriel-p 's review that would require me to review the paper or code again, it's all good for me!

If there is anything else I need to do, or I forgot something, let me know @xuanxu :smiley:

Merry Christmas to y'all :christmas_tree:
