### 2019-12-20 #1940 @Gabriel-p
Other than a very minor issue with the docs (just opened an issue at the code's repo), it looks good to me!
