### 2019-12-20 b437b8a
Docs Fix
Update indented code lines and moment map
