+++ docs/moment_maps.rst
@@ -18,9 +18,9 @@ moment maps with the survey::
     >>> ax = fig.add_subplot(111, projection = ccrs.Mollweide(central_longitude=0))
 
     >>> fig = survey.intensity_map(ax = ax, c = mean_vel, vmin = -75, vmax = 75, 
-                                   norm = None, cmap = "RdBu_r", colorbar = True, 
-                                   cbar_kwargs={"orientation":"horizontal", 
-                                                "label":"Mean Velocity (km/s)"})
+    ...                            norm = None, cmap = "RdBu_r", colorbar = True, 
+    ...                            cbar_kwargs={"orientation":"horizontal", 
+    ...                                         "label":"Mean Velocity (km/s)"})
 
 .. image:: images/moment1_map.png
    :width: 600
