+++ docs/stacking.rst
@@ -19,22 +19,22 @@ For example, we can stack the first 10 spectra in the survey::
     >>> fig = plt.figure()
     >>> ax = fig.add_subplot(111)
     >>> ax.plot(stacked_spectrum["VELOCITY"][0], 
-    >>>         stacked_spectrum["DATA"][0], 
-    >>>         color = 'r', 
-    >>>         lw = 2, 
-    >>>         alpha = 0.8, 
-    >>>         label = r"Stacked $H\alpha$")
+    ...         stacked_spectrum["DATA"][0], 
+    ...         color = 'r', 
+    ...         lw = 2, 
+    ...         alpha = 0.8, 
+    ...         label = r"Stacked $H\alpha$")
     >>> ax.fill_between(stacked_spectrum["VELOCITY"][0], 
-    >>>                 stacked_spectrum["CI"][0][0], 
-    >>>                 stacked_spectrum["CI"][0][1], 
-    >>>                 color = "r", 
-    >>>                 alpha = 0.1, 
-    >>>                 label = "95% Confidence Interval")
+    ...                 stacked_spectrum["CI"][0][0], 
+    ...                 stacked_spectrum["CI"][0][1], 
+    ...                 color = "r", 
+    ...                 alpha = 0.1, 
+    ...                 label = "95% Confidence Interval")
 
     >>> ax.set_xlabel("LSR Velocity (km/s)", fontsize = 12)
     >>> ax.set_ylabel("Intensity (Rayleighs / (km/s))", fontsize = 12)
     >>> ax.set_title("{0}".format(stacked_spectrum["NAME"][0]), 
-    >>>              fontsize = 12)
+    ...              fontsize = 12)
 
     >>> leg = ax.legend(fontsize = 12)
 
