### 2019-12-20 #1940 @Deech08
Thanks @HeloiseS and @Gabriel-p for all the help and feedback, and @xuanxu for coordinating the review process! I've made a few final changes to fix to the minor documentation issues and all issues on the code repo are closed. I do want to change the package version to v1.0.0 when all review is complete to go along with the paper release so I'm wondering when the best time to do that is?
