### 2019-12-23 984cee3
fix paper citation
add semicolon
