### 2019-12-23 #1940 @Deech08
Thanks again for your help!
I did notice one small issue in the paper - the additional astropy citation (2018) wasn't displaying properly, but is now fixed.

The final release version is v1.0.1 (on pypi and Github).
It is on Zenodo at the following:
https://doi.org/10.5281/zenodo.3591960
