### 2019-12-24 #1940 @arfon
@Gabriel-p, @HeloiseS - many thanks for your reviews and to @xuanxu for editing this submission ✨

@Deech08 - your paper is now accepted into JOSS :zap::rocket::boom:⛄️
